feat:  Created page for examining unused custom utility classes
The custom utility class is not included in the final bundle.
There are some side-effects when loading the class in the dev mode live server. Once the class is loaded through the code, when removing the class, the class is still in memory of the live server. The server needs to be restarted in order to forget about the unused class.